use long not int for SNF invariants

+++ pari_snf.h
@@ -7,7 +7,7 @@
 
 // Return a list of the Smith Normal Form invariants of a matrix
 // (encoded as vector<vector<int>>)
-vector<int> invariants(const vector<vector<int>>& M);
+vector<long> invariants(const vector<vector<int>>& M);
 
 // Return list of entries of the HNF of M (exncoded as above),
 // assuming M has size 2 (i.e. represnts a 2xn matrix):
